+++ src/pages/mentor/task/detail/ui/ImageSlider.tsx
